+++ docs/pages/Home.vue
@@ -6,6 +6,13 @@
             <hr>
             <h2 class="subtitle">Dialog</h2>
             <button class="button" @click="alertDialog">Alert</button>
+            <button class="button" @click="alertDialogWTitle">Alert with title</button>
+
+            <br><br>
+
+            <button class="button" @click="confirmDialog">Confirm</button>
+            <button class="button" @click="confirmDialogCustom">Confirm customized</button>
+            <button class="button" @click="confirmDialogCustomWCallback">Confirm customized with callback</button>
 
             <hr>
             <h2 class="subtitle">Snackbar</h2>
@@ -171,7 +178,7 @@
 
             <br><br>
 
-            <b-tooltip label="Maybe you like animation" position="is-right" animated>
+            <b-tooltip label="Maybe you like a little animation" position="is-right" animated>
                 <button class="button">Animated</button>
             </b-tooltip>
 
@@ -188,23 +195,6 @@
             <b-tooltip label="Or light?" type="is-light">
                 <button class="button">Light</button>
             </b-tooltip>
-
-            <b-tooltip label="Maybe green?" type="is-success">
-                <button class="button">Success</button>
-            </b-tooltip>
-
-            <b-tooltip label="... red?" type="is-danger">
-                <button class="button">Danger</button>
-            </b-tooltip>
-
-            <b-tooltip label="Blue!" type="is-info">
-                <button class="button">Info</button>
-            </b-tooltip>
-
-            <b-tooltip label="No, yellow!" type="is-warning">
-                <button class="button">Warning</button>
-            </b-tooltip>
-
         </div>
     </section>
 </template>
@@ -275,8 +265,42 @@
 
             alertDialog() {
                 this.$dialog.alert({
-                    title: 'Title Alert',
                     message: 'Everything looks <strong>fine</strong>!'
+                })
+            },
+            alertDialogWTitle() {
+                this.$dialog.alert({
+                    title: 'Title Alert',
+                    message: 'I have a title! And a custom button!',
+                    confirmText: 'Cool!'
+                })
+            },
+            confirmDialog() {
+                this.$dialog.confirm({
+                    message: 'Continue on this task?'
+                })
+            },
+            confirmDialogCustom() {
+                this.$dialog.confirm({
+                    title: 'Privacy Politics',
+                    message: ' Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce id fermentum quam. Proin sagittis, nibh id hendrerit imperdiet, elit sapien laoreet elit, ac scelerisque diam velit in nisl. Nunc maximus ex non laoreet semper. Nunc scelerisque, libero sit amet pretium dignissim, augue purus placerat justo, sit amet porttitor dui metus in nisl. Nulla non leo placerat, porta metus eu, laoreet risus. Etiam lacinia, purus eu luctus maximus, elit ex viverra tellus, sit amet sodales quam dui nec odio. Nullam porta mollis est. Quisque aliquet malesuada fringilla. Pellentesque volutpat lacus at ante posuere, non pulvinar ante porta. Proin viverra eu massa nec porta. Aliquam rhoncus velit quis sem hendrerit, ut dictum nisl accumsan. Maecenas erat enim, scelerisque non ligula ac, eleifend venenatis ligula. Praesent molestie mauris sed elit posuere, non malesuada libero gravida. In hac habitasse platea dictumst. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. ',
+                    cancelText: 'Disagree',
+                    confirmText: 'Agree',
+                    type: 'is-success'
+                })
+            },
+            confirmDialogCustomWCallback() {
+                this.$dialog.confirm({
+                    title: 'Deleting account',
+                    message: 'Are you sure you want to <strong>delete</strong> your account? This action cannot be undone.',
+                    confirmText: 'Delete Account',
+                    type: 'is-danger',
+                    showIcon: true,
+                    onConfirm: () => {
+                        this.$toast.open({
+                            message: 'Account deleted!'
+                        })
+                    }
                 })
             }
         }