+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 * New component: Message
 * New component: Tooltip
+* New component: Dialog
 * Added size option to Pagination
 * Added color option to Icon
 * Notification slightly modified to keep original Bulma style
+* Toasts and Snackbars now queue between each other
+* Toast duration reduced to 2.5s
 
 ## 0.0.2